Clean up CamanInstance's treatment of selectors vs IDs.
Previously the passed-in selector "#image-id" was used as the Store key,
and whether various id fields were set to "#image-id" or "image-id" was
not consistent. Now all id fields are set to "image-id", and Store strips
off the leading "#" for string search terms. Store works correctly when
mixing the selector and HTML object Caman() forms.